Add multiple runtimes benchmark
