+++ src/FlakeId.Benchmarks/Program.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using BenchmarkDotNet.Attributes;
-using BenchmarkDotNet.Diagnostics.Windows.Configs;
-using BenchmarkDotNet.Running;
-using IdGen;
-using MassTransit;
+﻿using BenchmarkDotNet.Running;
 
 namespace FlakeId.Benchmarks
 {
@@ -12,38 +7,7 @@ namespace FlakeId.Benchmarks
         private static void Main(string[] args)
         {
             BenchmarkRunner.Run<IdCreationBenchmarks>();
-        }
-    }
-
-    [DisassemblyDiagnoser]
-    [InliningDiagnoser(true, null)]
-    public class IdCreationBenchmarks
-    {
-        private static readonly IdGenerator s_idGenerator = new IdGenerator(10,
-            new IdGeneratorOptions(sequenceOverflowStrategy: SequenceOverflowStrategy.SpinWait));
-
-        [Benchmark]
-        public void Single_FlakeId()
-        {
-            Id.Create();
-        }
-
-        [Benchmark]
-        public void Single_Guid()
-        {
-            Guid.NewGuid();
-        }
-
-        [Benchmark]
-        public void Single_NewId()
-        {
-            NewId.Next();
-        }
-        
-        [Benchmark]
-        public void Single_IdGen()
-        {
-            s_idGenerator.CreateId();
+            BenchmarkRunner.Run<FlakeIdMultipleRuntimesBenchmarks>();
         }
     }
 }
